Tidy, remove second controller
Not committed to this approach though
