+++ src/ApprenticePortal/SFA.DAS.ApprenticePortal.Api/Controllers/ApprenticeHomepageController.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using SFA.DAS.ApprenticePortal.Application.ApprenticeAccounts.Queries;
+using SFA.DAS.ApprenticePortal.Application.Commands.ApprenticeUpdate;
 using SFA.DAS.ApprenticePortal.Application.Homepage.Queries;
+using SFA.DAS.ApprenticePortal.InnerApi.ApprenticeAccounts.Requests;
 using System;
 using System.Threading.Tasks;
 
@@ -24,6 +26,36 @@ namespace SFA.DAS.ApprenticePortal.Api.Controllers
                 return NotFound();            
 
             return Ok(result.ApprenticeHomepage);
+        }
+
+        [HttpGet]
+        [Route("/apprentices/{id}")]
+        public async Task<IActionResult> GetApprentice(Guid id)
+        {
+            var queryResult = await _mediator.Send(new GetApprenticeQuery
+            {
+                ApprenticeId = id
+            });
+
+            if (queryResult.Apprentice == null)
+                return NotFound();
+
+            return Ok(queryResult.Apprentice);
+        }
+
+        [HttpPatch("/apprentices/{apprenticeId}")]
+        public async Task<IActionResult> UpdateApprentice(Guid apprenticeId, UpdateApprenticeRequest request)
+        {
+            await _mediator.Send(new ApprenticeUpdateCommand
+            {
+                ApprenticeId = request.ApprenticeId,
+                FirstName = request.FirstName,
+                LastName = request.LastName,
+                Email = request.Email,
+                DateOfBirth = request.DateOfBirth
+            });
+
+            return NoContent();
         }
     }
 }
